+++ src/communications/communication.ts
@@ -45,8 +45,6 @@ export abstract class Communication<ConnectionData> {
     abstract info(): Promise<FwcAgentInfo>;
 
     abstract installPlugin(name: string,enabled: boolean): Promise<string>;
-    
-    abstract createWebSocket(): Promise<string>;
 
     protected handleRequestException(error: Error, eventEmitter?: EventEmitter) {
         if (errorHasCode(error)) {