+++ src/communications/ssh.communication.ts
@@ -195,8 +195,4 @@ export class SSHCommunication extends Communication<SSHConnectionData> {
     installPlugin(name: string,enabled: boolean): Promise<string> {
         throw new Error("Method not implemented.");
     }
-
-    createWebSocket(): Promise<string> {
-        throw new Error("Method not implemented.");
-    }
 }